+++ packages/@expo/cli/CHANGELOG.md
@@ -25,6 +25,7 @@
 ### 💡 Others
 
 - Export `_expo/.routes.json` metadata file for `static` outputs ([#40243](https://github.com/expo/expo/pull/40243) by [@kitten](https://github.com/kitten))
+- Add environmental variable to enable experimental web modal implementation ([#40231](https://github.com/expo/expo/pull/40231) by [@Ubax](https://github.com/Ubax))
 
 ## 54.0.10 — 2025-10-01